requirements.txt
Pillow
django-extensions
django-robots
django-debug-toolbar
django-taggit